Refactor some duplicate methods/computed properties into mixins
