+++ mixins/mixinImages.js
@@ -1,0 +1,7 @@
+export default {
+  computed: {
+    hasImages() {
+      return this.images.length > 0
+    }
+  }
+}
